edycja maila w dyplomach

+++ app/application/models/Files.php
@@ -73,10 +73,17 @@ class Application_Model_Files extends GP_Application_Model
 	public function delete()
 	{
 		parent::delete();
-		unlink(APPLICATION_PATH.'/../public/assets/applications/'.$this->path);
+		if (file_exists(APPLICATION_PATH.'/../public/assets/applications/'.$this->path))
+		{
+			unlink(APPLICATION_PATH.'/../public/assets/applications/'.$this->path);
+		}
+			
 		foreach($this->_imageData as $key => $data)
 		{
-			unlink(APPLICATION_PATH.'/../public/assets/applications/'.$this->getImage($key));
+			if (file_exists(APPLICATION_PATH.'/../public/assets/applications/'.$this->getImage($key)))
+			{
+				unlink(APPLICATION_PATH.'/../public/assets/applications/'.$this->getImage($key));
+			}
 		}
 		
 		return $this;
